+++ src/rules/mod.rs
@@ -229,6 +229,37 @@ fn verify_no_rule_properties(properties: &RuleProperties) -> Result<(), RuleConf
     Ok(())
 }
 
+fn verify_required_properties(
+    properties: &RuleProperties,
+    names: &[&str],
+) -> Result<(), RuleConfigurationError> {
+    for name in names.iter() {
+        if !properties.contains_key(*name) {
+            return Err(RuleConfigurationError::MissingProperty(name.to_string()));
+        }
+    }
+    Ok(())
+}
+
+fn verify_property_collisions(
+    properties: &RuleProperties,
+    names: &[&str],
+) -> Result<(), RuleConfigurationError> {
+    let mut exists = false;
+    for name in names.iter() {
+        if properties.contains_key(*name) {
+            if exists {
+                return Err(RuleConfigurationError::PropertyCollision(
+                    names.iter().map(ToString::to_string).collect(),
+                ));
+            } else {
+                exists = true;
+            }
+        }
+    }
+    Ok(())
+}
+
 #[cfg(test)]
 mod test {
     use super::*;